Fixed up a whole pile of errors to make layout work
There was a bunch of problems here including missing </div> tags in the html file.

The main issue was that you were confused about the id of the html element vs the grid-area name in the css.

Also there were typos in the grid-area names, and missing grid-areas.

Note that it still isn't exactly matching the design, because of the different way Figma deals with different-sized images.  You will probably want to pre-process your images to all have the same or similar aspect ratios.